Refactor .clib schema to use a single set of tables instead of parallel tables for proteomics and small molecule
