+++ src/org/labkey/targetedms/chromlib/BaseDaoImpl.java
@@ -275,19 +275,4 @@ public abstract class BaseDaoImpl<T extends AbstractLibEntity> implements Dao<T>
     protected abstract void setValuesInStatement(T t, PreparedStatement stmt) throws SQLException;
 
     protected abstract ColumnDef[] getColumns();
-
-    protected int setIonMobilityColumns(PreparedStatement stmt, int colIndex, AbstractLibPrecursor<?> precursor) throws SQLException
-    {
-        stmt.setObject(colIndex++, precursor.getExplicitIonMobility(), Types.DOUBLE);
-        stmt.setObject(colIndex++, precursor.getCcs(), Types.DOUBLE);
-        stmt.setObject(colIndex++, precursor.getIonMobilityMS1(), Types.DOUBLE);
-        stmt.setObject(colIndex++, precursor.getIonMobilityFragment(), Types.DOUBLE);
-        stmt.setObject(colIndex++, precursor.getIonMobilityWindow(), Types.DOUBLE);
-        stmt.setString(colIndex++, precursor.getIonMobilityType());
-        stmt.setString(colIndex++, precursor.getExplicitIonMobilityUnits());
-        stmt.setObject(colIndex++, precursor.getExplicitCcsSqa(), Types.DOUBLE);
-        stmt.setObject(colIndex++, precursor.getExplicitCompensationVoltage(), Types.DOUBLE);
-        stmt.setObject(colIndex++, precursor.getPrecursorConcentration(), Types.DOUBLE);
-        return colIndex;
-    }
 }
