+++ src/org/labkey/targetedms/chromlib/LibInfo.java
@@ -34,9 +34,6 @@ public class LibInfo
     private int _precursors;
     private int _transitions;
 
-    private int _moleculeLists;
-    private int _molecules;
-
     public String getPanoramaServer()
     {
         return _panoramaServer;
@@ -127,26 +124,6 @@ public class LibInfo
         _transitions = transitions;
     }
 
-    public int getMolecules()
-    {
-        return _molecules;
-    }
-
-    public void setMolecules(int molecules)
-    {
-        _molecules = molecules;
-    }
-
-    public int getMoleculeLists()
-    {
-        return _moleculeLists;
-    }
-
-    public void setMoleculeLists(int moleculeLists)
-    {
-        _moleculeLists = moleculeLists;
-    }
-
     @Override
     public boolean equals(Object o)
     {
@@ -160,8 +137,6 @@ public class LibInfo
         if (_precursors != libInfo._precursors) return false;
         if (_proteins != libInfo._proteins) return false;
         if (_transitions != libInfo._transitions) return false;
-        if (_molecules != libInfo._molecules) return false;
-        if (_moleculeLists != libInfo._moleculeLists) return false;
         if (_container != null ? !_container.equals(libInfo._container) : libInfo._container != null) return false;
         if (_created != null ? !_created.equals(libInfo._created) : libInfo._created != null) return false;
         if (_panoramaServer != null ? !_panoramaServer.equals(libInfo._panoramaServer) : libInfo._panoramaServer != null)
@@ -184,8 +159,6 @@ public class LibInfo
         result = 31 * result + _peptides;
         result = 31 * result + _precursors;
         result = 31 * result + _transitions;
-        result = 31 * result + _molecules;
-        result = 31 * result + _moleculeLists;
         return result;
     }
 }
